pyproject.toml
[project]
name = "hello-lambda-cdk"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
